+++ apps/desktop/src-tauri/src/ai/client_real_anthropic_test.rs
@@ -75,7 +75,7 @@ async fn smoke_claude_haiku_stream() {
     .expect("stream open");
 
     let mut text = String::new();
-    let mut chunks: usize = 0;
+    let mut chunks: u64 = 0;
     while let Some(item) = stream.next().await {
         let chunk = item.expect("chunk ok");
         text.push_str(&chunk);
